server/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
alpaca-trade-api
plaid-python
